+++ internal/generic/thrift/base.go
