+++ internal/generic/thrift/http.go
