+++ internal/generic/thrift/http_fallback.go
