+++ internal/generic/thrift/http_go116plus_amd64.go
