+++ internal/generic/thrift/json_go116plus_amd64.go
