+++ internal/generic/thrift/struct.go
