+++ internal/generic/thrift/thrift.go
