+++ internal/generic/thrift/util.go
