+++ internal/generic/thrift/write.go
