+++ internal/generic/thrift/write_test.go
