+++ pkg/generic/generic_service_test.go
@@ -24,9 +24,9 @@ import (
 
 	"github.com/golang/mock/gomock"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	mocks "github.com/cloudwego/kitex/internal/mocks/generic"
 	"github.com/cloudwego/kitex/internal/test"
-	gthrift "github.com/cloudwego/kitex/pkg/generic/thrift"
 	"github.com/cloudwego/kitex/pkg/remote"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
 	"github.com/cloudwego/kitex/pkg/utils"
@@ -80,7 +80,7 @@ func TestGenericService(t *testing.T) {
 	err = r.Write(ctx, method, buffer)
 	test.Assert(t, err.Error() == "unexpected Result writer type: <nil>")
 	// Write expect
-	resultWriteInner.EXPECT().Write(ctx, buffer, r.Success, (*gthrift.Base)(nil)).Return(nil).AnyTimes()
+	resultWriteInner.EXPECT().Write(ctx, buffer, r.Success, (*thrift.Base)(nil)).Return(nil).AnyTimes()
 	r.SetCodec(resultWriteInner)
 	// write ok
 	err = r.Write(ctx, method, buffer)
