+++ pkg/generic/httppbthrift_codec_test.go
@@ -24,8 +24,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	gthrift "github.com/cloudwego/kitex/pkg/generic/thrift"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
 )
 
@@ -70,8 +70,8 @@ func TestHTTPPbThriftCodec(t *testing.T) {
 	test.Assert(t, htc.svcName == "ExampleService")
 
 	rw := htc.getMessageReaderWriter()
-	_, ok := rw.(gthrift.MessageWriter)
+	_, ok := rw.(thrift.MessageWriter)
 	test.Assert(t, ok)
-	_, ok = rw.(gthrift.MessageReader)
+	_, ok = rw.(thrift.MessageReader)
 	test.Assert(t, ok)
 }
