+++ pkg/generic/httpthrift_codec_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/bytedance/sonic"
 	"github.com/cloudwego/dynamicgo/conv"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	gthrift "github.com/cloudwego/kitex/pkg/generic/thrift"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
 )
 
@@ -71,9 +71,9 @@ func TestHttpThriftCodec(t *testing.T) {
 	test.Assert(t, !ok)
 
 	rw = htc.getMessageReaderWriter()
-	_, ok = rw.(gthrift.MessageWriter)
+	_, ok = rw.(thrift.MessageWriter)
 	test.Assert(t, ok)
-	_, ok = rw.(gthrift.MessageReader)
+	_, ok = rw.(thrift.MessageReader)
 	test.Assert(t, ok)
 }
 
@@ -103,9 +103,9 @@ func TestHttpThriftCodecWithDynamicGo(t *testing.T) {
 	test.Assert(t, htc.svcName == "ExampleService")
 
 	rw := htc.getMessageReaderWriter()
-	_, ok := rw.(gthrift.MessageWriter)
+	_, ok := rw.(thrift.MessageWriter)
 	test.Assert(t, ok)
-	_, ok = rw.(gthrift.MessageReader)
+	_, ok = rw.(thrift.MessageReader)
 	test.Assert(t, ok)
 }
