+++ pkg/generic/jsonthrift_codec_test.go
@@ -21,8 +21,8 @@ import (
 
 	"github.com/cloudwego/dynamicgo/conv"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	"github.com/cloudwego/kitex/pkg/generic/thrift"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
 )
