+++ pkg/generic/mapthrift_codec_test.go
@@ -19,8 +19,8 @@ package generic
 import (
 	"testing"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	"github.com/cloudwego/kitex/pkg/generic/thrift"
 	"github.com/cloudwego/kitex/pkg/serviceinfo"
 )
