+++ pkg/generic/thriftidl_provider_test.go
@@ -22,8 +22,8 @@ import (
 
 	dthrift "github.com/cloudwego/dynamicgo/thrift"
 
+	"github.com/cloudwego/kitex/internal/generic/thrift"
 	"github.com/cloudwego/kitex/internal/test"
-	"github.com/cloudwego/kitex/pkg/generic/thrift"
 )
 
 func TestAbsPath(t *testing.T) {
